workshop/requirements_additions.txt
pydantic[email]>=2.8
python-dotenv>=1.0
openai>=1.51
pyyaml>=6.0.1
